refactor: migrate setup creation suites to scenario tables (#4273)
Densify four near-cap game_setup_creation_and_assignment_* suites into
thin runners plus shared scenario tables, completing slice D migration
for creation/assignment coverage.

Refs #4273
